Extract separate application module
- Allows root module to have a public API
